backend-python/requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PaperQA2 和 Docling
paper-qa>=5.0.0,<6.0.0
docling>=2.0.0,<3.0.0

# LLM相关
litellm>=1.0.0,<2.0.0

# 向量数据库
pgvector>=0.2.0
psycopg[binary]>=3.1.0
asyncpg>=0.29.0
sentence-transformers>=2.0.0,<3.0.0

# 图数据库
neo4j>=5.14.0

# 全文搜索
tantivy>=0.20.0

# 工具库
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.0.0

# 日志
structlog>=23.0.0

# 缓存
redis>=5.0.0,<6.0.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0
black>=23.0.0
isort>=5.12.0

# JWT认证 (内部服务通信)
pyjwt[crypto]==2.8.0
cryptography==42.0.5
